refactor: general improvements (#81)
* fix: use correct shield for issues by label

* style: run `deno fmt`

* build(CODEOWNERS): remove extra `#` and `\n`

* docs: remove escaping `.` for folders

* docs: commit to using `userstyle`

Originally, we wanted to avoid using the
word userstyle in case non-developers
were confused.

However, given that the name of the
repository is catppuccin/userstyles and
there are now explanations for what a
userstyle is. I believe we should be more
consistent and commit to using the word.

* build(userstyle.yml): shorten issue template

This commit uses the abbreviated YAML list
syntax to auto-populate the labels. This will
ensure that this file remains easy to scroll
as more userstyles are added.

* chore: use raw link for schema

Using a relative link only works when the
user has locally cloned the repository. This
commit changes it back to a raw link so
that it overcomes this limitation.

* build: refine commit msg in `git-auto-commit-action"

The current commit message is quite verbose and
co-authors the original author.

In @nekowinston's words:
"there's no reason i'm co-authored when i merge
a PR in that changes a single line in the readme
:baldkek:"

This commit reduces the verbosity of the commit
message and doesn't co-author the merger.

* build(issue-labeler.yml): shorten file

Please see the commit description of
4a1326983f1e1a3bbe663f4aa16b6e19f8df8da8

* docs: highlight text as bold

We want to draw the user's attention
to specific parts of the page.

For the CONTRIBUTING.md, we want to
make sure their eyes are down towards
"New Userstyles."

For the userstyle-creation.md, we want
to make sure that the user knows READMEs
are auto-generated.

* chore: update `deno.lock`

* chore: generate health files