fix(reddit): theme credit bar background
Fixes the pdp-credit-bar element background that shows the subreddit name, OP
username, and time since it was posted (e.g. 3h ago) which was displaying
with --color-neutral-background instead of matching the @base color.

For example, on Latte, it would appear lighter than base color. On any of
the dark flavors, it would appear darker than base color.